chore: add `.dev.vars.example` file and update `Dockerfile`
This commit introduces a new .dev.vars.example file, which includes environment variables necessary for the proper functioning of the application. Additionally, the Dockerfile has been updated to copy the .dev.vars.example file into the docker container. This change is important to ensure that all the required environment variable templates are correctly set up and available in the container environment during runtime.